Sketch out a local run
