+++ ingestion-lambda/package.json
@@ -3,6 +3,7 @@
 	"version": "1.0.0",
 	"description": "",
 	"scripts": {
+		"local": "tsx watch localRun.ts",
 		"dev": "INGESTION_LAMBDA_QUEUE_URL=http://sqs.eu-west-1.localhost.localstack.cloud:4566/000000000000/local-ingestion-queue tsx watch localRun.ts",
 		"db-refresh": "tsx watch randomisedDatabaseInserts.ts",
 		"typecheck": "tsc -noEmit",
